requirements.txt
tkinterweb
screeninfo
pywin32
pandas
python-socketio
requests
openpyxl